Extricates main staker control panel to component; Code Reorg.
